YUK-18525 : CapBank Default Reclose Delay Should be 3 or 5 Minutes

+++ yukon-client/common/src/com/cannontech/database/db/capcontrol/CapBank.java
@@ -17,7 +17,7 @@ public class CapBank extends com.cannontech.database.db.DBPersistent
 	private String typeOfSwitch = StringUtils.EMPTY;
 	private String switchManufacture = StringUtils.EMPTY;
 	private String mapLocationID = "0";  //old integer default
-	private Integer recloseDelay = new Integer(0);
+	private Integer recloseDelay = new Integer(300);
 	private Integer maxDailyOps = new Integer(0);
 	private Character maxOpDisable = new Character('N');
